Rewrite A2 script so that the function performing the execution e.g. execute of the sql statement (update) is called from a function that forms an update statement
E.g.: update, and returns the result of the execution of the sql statement.

The function to execute the sql statement returns the number of lines affected by the statement.

To avoid memory leaks close the cursor before returning from the function. Use contextlib closing for managing the cursor. It i possible to have several context managers in one with statement