Rewrite A4 script so that the opening and closing of the connection to the database is handled in one function e.g. main using a separate class to handle the context management.
The class defines __enter__  method which, in case there was an exception when opening the connection, calls a method that handles the exception and calls __exit__. Other exceptions raised when the connection is open are handled in __exit__ method prior to closing the connection. If the exception type is sqlite3 error (test: issubclass(exc_type, sqlite3.Error), the exception is handled and the method returns True. Other exceptions are not handled but passed on and the return value is False.